cont

+++ lib/movies/data/models/movie_model.dart
@@ -16,7 +16,7 @@ class MovieModel extends Movie {
         title: json['title'],
         backdropPath: json['backdrop_path'],
         overview: json['overview'],
-        voteAverage: json["vote_average"].toDouble(),
+        voteAverage: json["vote_average"],
         genreIds: List<int>.from(json['genre_ids'].map((e) => e)),
         releaseDate: json['release_date'],
       );
